huge refactoring
Now little better setup and now I feel even more that I switch to
LazyVim and bite the dust to migrate my own changes.

There is a lot to do still, and it seems that fixing this is bigger task
than migrating to LazyVim.

Personally I don't really like that much pre-made bundles of configs
that much, but to keep up with my config and things I want here starts
to be too huge amount to keep up myself.

Most of the things in LazyVim what I would need to change is some
keybindings and add org-mode there, and perhaps some small things.

I have used lazy previously but went away for a while.